Move the actual runner into a new Elm file that is included at runtime.
This is predominantly because there's parts of the internals that are
necessarily included in the API. Changing `TestResult`, which was never
supposed to be exposed, has meant that I will have to bump this to
v2.0.0 to publish it. Making sure the runner only gets built in later
means that we can keep these details private and just worry about the
`Test` and `Assertion` interfaces.